Include vendor for PHP7x,8.0,<8.1.14 and PHP >= 8.1.14.
There's a switch that includes either one of the vendor
dir, depending on the PHP version running at the system. The reason
is that the Yaml library would not work correctly with newer PHP
versions. And update of the Yaml library dropped support for PHP < 8.1
but Moodle still supports it. That made the two versions necessary.