make detail rating comment fields work with all editors
In the detail ratings the student and private comments
could be edited with atto and tiny. However, when clicking
one of the rating tags (--, -, +, ++) the atto only was filled
with the text block. This has been changed to work with
tinyMCE as well. Furthermore, if the user set in his preferences
to use the plain text area, this didn't work at all.